feat: 传递并处理array

+++ cpp/include/cpp_number.hpp
@@ -5,7 +5,7 @@
 #include <string>
 
 int cpp_process_number(
-    const usize rust_int,
+    const int32_t rust_int,
     const float rust_float,
     const double rust_double
 );
@@ -16,9 +16,9 @@ extern "C"
 #endif
 
     int trans_number(
-        const int rust_int,
-        const float rust_float,
-        const double rust_double
+        const int32_t rust_int,
+        const float rust_float,  // 32位
+        const double rust_double // 64位
     );
 
 #ifdef __cplusplus
